requirements.txt
torch==2.11.0+cu128
torchvision==0.26.0
transformers==5.6.2
Pillow==12.2.0
numpy==2.4.4
timm==1.0.26
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.12.0
tqdm==4.67.3
sentencepiece>=0.2.0
protobuf>=4.25.0
